Adding FashionMNIST dataset code
Fashion MNIST Dataset with custom labels
This is an image dataset based on publicly available Fashion MNIST Dataset dataset of small (28x28 pixels), grey-scale images of clothing items such as shoes, coats, pants, etc. The dataset was create to serve as a direct drop-in replacement for the well known MNIST dataset for benchmarking machine learning algorithms. It shares the same image size and structure. The target field was originally assigned a label for clothing type to each image as an integer from 0-9. The training set contains 60,000 examples and the test set 10,000 examples.

For this assignment, we are providing you the same input features, but the labels you will be using will be new. We have created a new label based on the data and calculated to be associated with each entry. The target is the modified label for the images obtained based on a formulation to categorize the fashion MNIST images to a new set of categories. The new categories are numbered 1-5, however, the detail regarding the categories is held out on purpose. The training dataset that you download from LEARN under Assignment 4 will have the "FashionMNIST with a Twist" dataset using these new labels.

The classification label is the feature target.